Simplified the Card class
Attempting to move away from mapping to arbitrary numbers, just to see if it makes for a cleaner implementation.  Using array indicies where necessary, but trying to use the very readable character representations wherever possible.  The eventual goal is to never see or care about the backing array index values.

Fixed a couple of unit tests that this breaks.  Many more tests are still broken.